+++ Backend/src/routes.ts
@@ -1,12 +1,13 @@
 import { Router } from 'express';
 
 import accountsRoutes from '@modules/accounts/account.routes.js';
+import attachmentRoutes from '@modules/attachments/attachment.routes.js';
 import demoRoutes from '@modules/demo/demo.routes.js';
+import draftRoutes from '@modules/drafts/draft.routes.js';
 import emailsRoutes from '@modules/emails/email.routes.js';
 import foldersRoutes from '@modules/folders/folder.routes.js';
 import usersRoutes from '@modules/user/user.routes.js';
 import utilsRoutes from '@modules/utils/index.js';
-import attachmentRoutes from '@modules/attachments/attachment.routes.js';
 
 const router = Router();
 
@@ -25,6 +26,8 @@ router.use('/emails', emailsRoutes);
 router.use('/folders', foldersRoutes);
 
 router.use('/attachments', attachmentRoutes);
+
+router.use('/drafts', draftRoutes);
 
 router.use('/utils', utilsRoutes);